agents/supplier-agent/requirements.txt
google-cloud-alloydb-connector[pg8000]
python-dotenv
google-genai
a2a-sdk[http-server]==0.3.2
uvicorn
starlette
requests
google-auth
